+++ openfeature-provider/java/src/main/java/com/spotify/confidence/RemoteResolver.java
@@ -4,6 +4,8 @@ import com.spotify.confidence.flags.resolver.v1.ResolveFlagsRequest;
 import com.spotify.confidence.flags.resolver.v1.ResolveFlagsResponse;
 import java.util.concurrent.CompletableFuture;
 
-public non-sealed interface ResolverFallback extends StickyResolveStrategy {
+public interface RemoteResolver {
   CompletableFuture<ResolveFlagsResponse> resolve(ResolveFlagsRequest request);
+
+  void close();
 }